--- src/main/java/com/kakaotech/team18/backend_server/application/Status.java
@@ -1,7 +1,0 @@
-package com.kakaotech.team18.backend_server.application;
-
-public enum Status {
-    PENDING,
-    PASS,
-    FAIL,
-}
