--- src/main/java/com/kakaotech/team18/backend_server/applicationFormField/FieldType.java
@@ -1,7 +1,0 @@
-package com.kakaotech.team18.backend_server.applicationFormField;
-
-public enum FieldType {
-    TEXT,
-    RADIO,
-    CHECKBOX
-}
