--- src/main/java/com/kakaotech/team18/backend_server/clubAdmin/Role.java
@@ -1,7 +1,0 @@
-package com.kakaotech.team18.backend_server.clubAdmin;
-
-public enum Role {
-    ADMIN,
-    MEMBER,
-    CLUB_ADMIN
-}
