--- src/main/java/com/kakaotech/team18/backend_server/clubMember/ActiveStatus.java
@@ -1,6 +1,0 @@
-package com.kakaotech.team18.backend_server.clubMember;
-
-public enum ActiveStatus {
-    ACTIVATE,
-    DEACTIVATE
-}
