+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationFormAnswer/ApplicationFormAnswer.java
@@ -1,7 +1,7 @@
-package com.kakaotech.team18.backend_server.applicationForm;
+package com.kakaotech.team18.backend_server.domain.applicationFormAnswer;
 
-import com.kakaotech.team18.backend_server.application.Application;
-import com.kakaotech.team18.backend_server.club.Club;
+import com.kakaotech.team18.backend_server.domain.application.Application;
+import com.kakaotech.team18.backend_server.domain.applicationForm.ApplicationForm;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -9,28 +9,26 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApplicationForm {
+public class ApplicationFormAnswer {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "application_form_id")
+    @Column(name = "application_form_answer_id")
     private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "club_id")
-    private Club club;
-
-    @ManyToOne(fetch = FetchType.LAZY)
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "application_id")
     private Application application;
 
-    private String title;
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "application_form_id")
+    private ApplicationForm applicationForm;
 
-    private String description;
+    private String answer;
 }
