+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationFormField/ApplicationFormField.java
@@ -1,34 +1,36 @@
-package com.kakaotech.team18.backend_server.applicationFormAnswer;
+package com.kakaotech.team18.backend_server.domain.applicationFormField;
 
-import com.kakaotech.team18.backend_server.application.Application;
-import com.kakaotech.team18.backend_server.applicationForm.ApplicationForm;
+import com.kakaotech.team18.backend_server.domain.applicationForm.ApplicationForm;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ApplicationFormAnswer {
+public class ApplicationFormField {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "application_form_answer_id")
+    @Column(name = "application_form")
     private Long id;
 
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "application_id")
-    private Application application;
-
-    @OneToOne(fetch = FetchType.LAZY)
+    @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "application_form_id")
     private ApplicationForm applicationForm;
 
-    private String answer;
+    private String question;
+
+    @Enumerated(EnumType.STRING)
+    private FieldType fieldType;
+
+    private String isRequired;
 }
