+++ src/main/java/com/kakaotech/team18/backend_server/domain/applicationFormField/FieldType.java
@@ -1,0 +1,7 @@
+package com.kakaotech.team18.backend_server.domain.applicationFormField;
+
+public enum FieldType {
+    TEXT,
+    RADIO,
+    CHECKBOX
+}
