+++ src/main/java/com/kakaotech/team18/backend_server/domain/clubMember/ActiveStatus.java
@@ -1,0 +1,6 @@
+package com.kakaotech.team18.backend_server.domain.clubMember;
+
+public enum ActiveStatus {
+    ACTIVATE,
+    DEACTIVATE
+}
