+++ src/main/java/com/kakaotech/team18/backend_server/domain/notification/Notification.java
@@ -1,11 +1,8 @@
-package com.kakaotech.team18.backend_server.clubMember;
+package com.kakaotech.team18.backend_server.domain.notification;
 
-import com.kakaotech.team18.backend_server.club.Club;
-import com.kakaotech.team18.backend_server.user.Users;
+import com.kakaotech.team18.backend_server.domain.club.Club;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -17,26 +14,18 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ClubMember {
+public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "club_member_id")
+    @Column(name = "notification_id")
     private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private Users user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "club_id")
     private Club club;
 
-    @Enumerated(EnumType.STRING)
-    private Position position;
+    private String title;
 
-    @Enumerated(EnumType.STRING)
-    private ActiveStatus activeStatus;
-
-
+    private String content;
 }
